+++ lib/mongo/auth/x509.ex
@@ -2,7 +2,7 @@ defmodule Mongo.Auth.X509 do
   @moduledoc false
   alias Mongo.MongoDBConnection.Utils
 
-  def auth({username, _password}, s) do
+  def auth({username, _password}, _db, s) do
     cmd = [authenticate: 1, user: username, mechanism: "MONGODB-X509"]
     with {:ok, _message} <- Utils.command(-2, cmd, s) do
       :ok